support src blocks that take time to return
I couldn't find a reliable Elisp API to wait that the comint process
ended (because I thing comint-send-input or gore split the input and
so it looks like the process returns multiple times). So I decided to
split the input myself and clean and aggregate the result myself. It
seems to work fine, although the waiting time is a bit long.